+++ ksef-client/src/main/java/pl/akmf/ksef/sdk/client/model/certificate/publickey/PublicKeyCertificate.java
@@ -7,6 +7,8 @@ import java.util.List;
 
 public class PublicKeyCertificate {
     private String certificate;
+    private String certificateId;
+    private String publicKeyId;
     private OffsetDateTime validFrom;
     private OffsetDateTime validTo;
     private List<PublicKeyCertificateUsage> usage = new ArrayList<>();
@@ -44,5 +46,21 @@ public class PublicKeyCertificate {
 
     public void setUsage(List<PublicKeyCertificateUsage> usage) {
         this.usage = usage;
+    }
+
+    public String getCertificateId() {
+        return certificateId;
+    }
+
+    public void setCertificateId(String certificateId) {
+        this.certificateId = certificateId;
+    }
+
+    public String getPublicKeyId() {
+        return publicKeyId;
+    }
+
+    public void setPublicKeyId(String publicKeyId) {
+        this.publicKeyId = publicKeyId;
     }
 }